Proper scaling

+++ addons/v4pserial/v4pserial.h
@@ -8,6 +8,6 @@ V4pPolygonP v4p_decodePoints(V4pPolygonP p, char* s, int scale);
 V4pPolygonP v4p_quickPolygon(V4pProps t, V4pColor col, V4pLayer pz, char* s, int scale);
 V4pPolygonP v4p_decodePolygon(char* s, int scale);
 char* v4p_encodePolygon(V4pPolygonP p, int scale);
-V4pPolygonP v4p_decodeSVGPath(V4pPolygonP p, char* path, int scale);
+V4pPolygonP v4p_decodeSVGPath(V4pPolygonP p, char* path, float scale);
 
 #endif
